Added better documentation for `disable-unambiguous-requires`
